Created the packages of the MVC and renamed the model to searchLogic
