+++ src/Scrabble/searchLogic/CacheReplacementPolicy.java
@@ -1,4 +1,4 @@
-package Scrabble.Model;
+package Scrabble.searchLogic;
 
 public interface CacheReplacementPolicy {
     void add(String word);
