+++ src/Scrabble/searchLogic/IOSearcher.java
@@ -1,4 +1,4 @@
-package Scrabble.Model;
+package Scrabble.searchLogic;
 
 import java.nio.file.Files;
 import java.nio.file.Paths;
